// src/shared/web/api-types.d.ts
declare const appRouter: _trpc_server.TRPCBuiltRouter<{
// ...
}, _trpc_server.TRPCDecorateCreateRouterOptions<{
// ...
    account: _trpc_server.TRPCBuiltRouter<{
        ctx: Context;
        meta: object;
        errorShape: {
            message: string;
            data: {
                stack: undefined;
                fieldErrors: {} | undefined;
                details: Record<string, unknown> | undefined;
                code: _trpc_server.TRPC_ERROR_CODE_KEY;
                httpStatus: number;
                path?: string;
            };
            code: _trpc_server.TRPC_ERROR_CODE_NUMBER;
        };
        transformer: true;
    }, _trpc_server.TRPCDecorateCreateRouterOptions<{
        status: _trpc_server.TRPCQueryProcedure<{
            input: void;
            output: {
                name: string;
                city: string | null;
                hasPassword: boolean;
                hasSocial: boolean;
                isOrganizer: boolean;
                needsSetup: boolean;
            };
            meta: object;
        }>;
        completeSignUp: _trpc_server.TRPCMutationProcedure<{
            input: {
                name: string;
                accountType: "participant" | "organizer";
                password?: string | undefined;
                city?: string | null | undefined;
            };
            output: {
                organizerProfileId: string;
            } | {
                organizerProfileId: null;
            };
            meta: object;
        }>;
        setPassword: _trpc_server.TRPCMutationProcedure<{
            input: {
                password: string;
            };
            output: void;
            meta: object;
        }>;
        signInMethods: _trpc_server.TRPCQueryProcedure<{
            input: void;
            output: SignInMethods;
            meta: object;
        }>;
        unlinkProvider: _trpc_server.TRPCMutationProcedure<{
            input: {
                providerId: "google" | "facebook" | "apple";
            };
            output: void;
            meta: object;
        }>;
        privacy: _trpc_server.TRPCQueryProcedure<{
            input: void;
            output: {
                visibility: Visibility;
                attendanceVisibility: Visibility;
            };
            meta: object;
        }>;
        setPrivacy: _trpc_server.TRPCMutationProcedure<{
            input: {
                visibility: "private" | "public";
                attendanceVisibility: "private" | "public";
            };
            output: void;
            meta: object;
        }>;
        becomeOrganizer: _trpc_server.TRPCMutationProcedure<{
            input: void;
            output: {
                organizerProfileId: string;
            };
            meta: object;
        }>;
    }>>;
// ...
}>>;
type AppRouter = typeof appRouter;
// ...
type RouterOutputs = inferRouterOutputs<AppRouter>;